requirements.txt
sqlalchemy
requests
feedparser
beautifulsoup4
python-dateutil
pyyaml
flask